Dimensionless_Tech_Task/requirements.txt
Django==3.1.4
djangorestframework
opencv-python
numpy
matplotlib